+++ packages/svelte/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.0.0-beta.5
 
 * Add `hash`, `query`, and `state` props to `<Link>`, remove `details`.
 * Add `curiStore` function to automatically sync router and store. This can create a new store or update an existing store.